Started turn order system
Added rudimentary turn order system to the simulator, including a 'UnitActionPanel' in the EngagementWindow for testing purposes. 
Turn order functions approximately as in the games at first, but is subject to a minor desynchronization that seems to favor slower enemies. Fixes are still being researched (the most promising one that comes to mind is implementing the reserve system exactly as described in the step-by-step algorithm at the end of the turn order section in TFergusson's mechanics guide, in defiance of my own intuition). If no solution can be found, however, the algorithm as implemented in this commit is sufficient.