Established flow of actions within a turn
Rewrote UnitActionPanel (the old version is saved in OldUnitActionPanel) in anticipation of finally implementing actions. The panel now shows the basic actions a unit may take and includes dummied Move and Act commands; rules are in place to ensure all actions can only be used in the right order, cancelled at the right time, etc. The structure established here should stay in place as the proper functionality is developed for Move, Act and Wait.