Synchronized movement
Move and Wait functions are now functional. Upon using Wait, any enacted movement will be committed to the server and sent on to the other client and the results displayed on both screens.

One desired improvement for down the line: Teleportation through buildings with Fairy Shoes is not yet possible since non-walkable tiles are not included in the search for tiles within a unit's walking range. To fix this will require complicated changes in the way maps are constructed and interacted with. A task for down the line, once more fundamental concerns are taken care of.