Added win condition and room expiration
Now when all the units on one team reach 0 HP, the server announces the winner to the room. Petrified units are not yet counted as dead.

If someone leaves the room, the server announces their departure to the room. If everyone leaves a room, the server drops it from its game list. If one participant in an engagement leaves, the other can still take turns and possibly win the battle; this should be rectified before I get sidetracked by more glamorous features.

Note: I am now aware that all my server-side queues are totally not thread-safe, and require synchronize blocks to protect them. This should be amended quickly too.