Implemented directional targeting and fixed element resistance check
Added code for directional targeting to the highlightTargetableTiles() method in MapPanelTest (which still needs to be renamed), the absence of which was previously preventing attacks from being made with spears. Still figuring out the best way to extend this to infinite-range skills like Earth Render, but we'll cross that bridge when we get there.

Also fixed a particularly bug-ridden stretch of code in the main damage formula in FFTACalc that checked the target's equipment effects for any resistances or weaknesses to the attacker's element, and the attacker's equipment for any enhancing effects. ArrayIndexOutOfBoundsExceptions had been occurring every time a unit with an elemental weapon tried to target anyone with an attack. No longer; now all element-resisting/enhancing gear seems to be working properly, along with the Geomancy support ability.