Intermediate facing and fixed unarmed attacks
After an action has been committed, the acting unit now faces their target before the effect is processed.
Fixed a bug with unarmed strikes; there was this piece of code that seemed to add the right-hand weapon's power to the attacker's atk in the damage calculation if they're holding a shield, which seems blatantly wrong; I'm not sure if this piece of code has some forgotten purpose, or if it's always been there in error. It's gone now.